Labfiles/03-ai-agent-functions/Python/agent.py
import os
from dotenv import load_dotenv
from typing import Any

# Add references


# Create a function to submit a support ticket


def main(): 

    # Clear the console
    os.system('cls' if os.name=='nt' else 'clear')

    # Load environment variables from .env file
    load_dotenv()
    project_endpoint= os.getenv("PROJECT_ENDPOINT")
    model_deployment = os.getenv("MODEL_DEPLOYMENT_NAME")

    # Connect to the AI Project client
   

        # Create a FunctionTool definition
        

        # Initialize the agent with the FunctionTool
        

        # Create a thread for the chat session
       

        # Loop until the user types 'quit'
        while True:
            # Get input text
            user_prompt = input("Enter a prompt (or type 'quit' to exit): ")
            if user_prompt.lower() == "quit":
                break
            if len(user_prompt) == 0:
                print("Please enter a prompt.")
                continue

            # Send a prompt to the agent
            

            # Get the agent's response
            

            # Check the run status for failures


            # Process function calls
            

            # If there are function call outputs, send them back to the model
            

        # Clean up
        

if __name__ == '__main__': 
    main()
