Labfiles/03-ai-agent-functions/Python/requirements.txt
python-dotenv
azure-identity
azure-ai-projects>=2.0.0b1